refactor ealystop
