+++ bin/test.sh
@@ -1,0 +1,1 @@
+python -m unittest test.$1
