+++ utils/rpn_msr/config.py
@@ -25,8 +25,8 @@ class Config:
     RPN_BBOX_INSIDE_WEIGHTS = (0, 1.0, 0, 1.0) # (x,y,dx,dy),之前作者都写成了[1,1,1,1]，不对，只保留y和dy
     RPN_POSITIVE_WEIGHT = -1.0
 
-    RPN_PRE_NMS_TOP_N = 12000
-    RPN_POST_NMS_TOP_N = 1000
+    RPN_PRE_NMS_TOP_N = 12000 #12000
+    RPN_POST_NMS_TOP_N = 1000 #1000
     RPN_NMS_THRESH = 0.7
     RPN_MIN_SIZE = 8
